pyproject.toml
[tool.poetry]
name = "money-tracker"
version = "0.1.0"
description = "UYU / US exchange rate monitor that scraps data from itau"
authors = ["elias <eliasmasquil@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
requests-html = "^0.10.0"
termplotlib = "^0.3.1"
click = "^7.1.2"
python-dotenv = "^0.13.0"
rich = "^1.2.3"
pandas = "^1.0.4"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
ipython = "^7.15.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
